fix: fix site theme
