+++ src/render.js
@@ -2,11 +2,39 @@ try {
   var fs = require("fs");
   var { semantic } = require("./highlight");
   var { num2hanzi } = require("./hanzi2num");
-  var { BOOK_COLORS, RED, BLACK } = require("./meta");
   var parser = require("./parser");
 } catch (e) {}
 
 const FONT = "'I.Ming', 'Source Han Serif KR', 'Noto Serif CJK KR', serif"; //"Source Han Serif TC"
+const RED = "#E53";
+const BLACK = "#222";
+
+const BOOK_COLORS = {
+  ctrl: RED,
+  lop: BLACK,
+  name: BLACK,
+  cmp: BLACK,
+  decl: BLACK,
+  print: BLACK,
+  rassgn: BLACK,
+  ctnr: BLACK,
+  comment: RED,
+  type: BLACK,
+  call: BLACK,
+  assgn: BLACK,
+  discard: BLACK,
+  endl: BLACK,
+  ans: BLACK,
+  expr: BLACK,
+  op: BLACK,
+  not: BLACK,
+  operand: BLACK,
+  bool: BLACK,
+  data: "#666",
+  iden: "#248",
+  quot: BLACK,
+  num: "#872"
+};
 
 function render(fname, txt, { plotResult = false } = {}) {
   var svgs = [];
